pyproject.toml
[tool.poetry]
name = "project-1"
version = "0.1.0"
description = ""
authors = ["evgeniatimoshkina <tea302@yandex.ru>"]
readme = "README.md"
packages = [{include = "project_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
envparse = "^0.2.0"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
drf-spectacular = "^0.26.2"
signals = "^0.0.2"
docker-compose = "^1.29.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
